requirements.txt
absl-py==1.4.0
alabaster==0.7.13
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
Av1an==6.1.5
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.9.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
css-html-js-minify==2.5.5
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
docutils==0.18.1
executing==1.2.0
fastapi==0.95.0
flake8==6.0.0
flatbuffers==23.3.3
fonttools==4.39.3
furo==2023.5.20
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
ml-dtypes==0.1.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.5
numpydoc==1.5.0
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.0
pyparsing==3.0.9
pytest==7.3.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scenedetect==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sh==2.0.3
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==6.2.1
sphinx-basic-ng==1.0.0b1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.26.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
text-unidecode==1.3
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
